+++ tt_metal/impl/CMakeLists.txt
@@ -101,6 +101,8 @@ set(IMPL_SRC
     ${CMAKE_CURRENT_SOURCE_DIR}/tensor/tensor_apis.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/tensor/tensor_utils.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/tensor/host_tensor.cpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/tensor/mesh_tensor.cpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/tensor/legacy_view.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/tensor/spec/tensor_spec.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/tensor/spec/layout/alignment.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/tensor/spec/layout/layout.cpp
